合并 1896575230585aba59c3ca8e7668c604e1dfbc60

+++ src/server/database/PostgreSQL.ts
@@ -1,18 +1,19 @@
+import type * as pg from 'pg';
 import {IDatabase, IGameShortData} from './IDatabase';
-import {Game, Score} from '../Game';
-import {GameOptions} from '../GameOptions';
-import {GameId, ParticipantId} from '../../common/Types';
+import {IGame, Score} from '../IGame';
+import {GameOptions} from '../game/GameOptions';
+import {GameId, ParticipantId, PlayerId, isGameId, safeCast} from '../../common/Types';
 import {SerializedGame} from '../SerializedGame';
 import {User} from '../User';
 import {Timer} from '../../common/Timer';
-import {Pool, ClientConfig} from 'pg';
 import {daysAgoToSeconds} from './utils';
 import {GameIdLedger} from './IDatabase';
 import {UserRank} from '../../common/rank/RankManager';
+import {oneWayDifference} from '../../common/utils/utils';
+import {Color} from '../../common/Color';
 // import {Rating} from 'ts-trueskill';
 
 export class PostgreSQL implements IDatabase {
-  protected client: Pool;
   private databaseName: string | undefined = undefined; // Use this only for stats.
 
   protected statistics = {
@@ -21,9 +22,17 @@ export class PostgreSQL implements IDatabase {
     saveConflictUndoCount: 0,
     saveConflictNormalCount: 0,
   };
+  private _client: pg.Pool | undefined;
+
+  protected get client(): pg.Pool {
+    if (this._client === undefined) {
+      throw new Error('attempt to get client before intialized');
+    }
+    return this._client;
+  }
 
   constructor(
-    config: ClientConfig = {
+    private config: pg.ClientConfig = {
       connectionString: process.env.POSTGRES_HOST,
     }) {
     if (config.connectionString?.startsWith('postgres')) {
@@ -40,26 +49,27 @@ export class PostgreSQL implements IDatabase {
         console.warn(e);
       }
     }
-    // Configuration stats saved for
-    this.client = new Pool(config);
   }
 
   public async initialize(): Promise<void> {
+    const {Pool} = await import('pg');
+    this._client = new Pool(this.config);
     await this.client.query('CREATE TABLE IF NOT EXISTS games(game_id varchar, save_id integer, game text, status text default \'running\', createtime timestamp(0) default now(), prop text, PRIMARY KEY (game_id, save_id))');
     // 测试时开启下一行， 时间戳带毫秒
     // await this.client.query('CREATE TABLE IF NOT EXISTS games(game_id varchar, save_id integer, game text, status text default \'running\', createtime timestamp default now(), prop text, PRIMARY KEY (game_id, save_id))');
     await this.client.query('CREATE TABLE IF NOT EXISTS participants(game_id varchar, participants varchar[], PRIMARY KEY (game_id))');
     await this.client.query('CREATE TABLE IF NOT EXISTS game_results(game_id varchar not null, seed_game_id varchar, players integer, generations integer, game_options text, scores text,createtime timestamp(0) default now(), PRIMARY KEY (game_id))');
+    await this.client.query('CREATE TABLE IF NOT EXISTS completed_game(game_id varchar not null, completed_time timestamp default now(), PRIMARY KEY (game_id))');
 
 
     await this.client.query('CREATE INDEX IF NOT EXISTS games_i1 on games(save_id)');
     await this.client.query('CREATE INDEX IF NOT EXISTS games_i2 on games(createtime)');
     await this.client.query('CREATE INDEX IF NOT EXISTS participants_idx_ids on participants USING GIN (participants)');
-
+    await this.client.query('CREATE INDEX IF NOT EXISTS completed_game_idx_completed_time on completed_game(completed_time)');
     await this.client.query('CREATE TABLE IF NOT EXISTS users(id varchar not null, name varchar not null, password varchar not null, prop varchar, createtime timestamp(0) default now(), PRIMARY KEY (id))');
 
     // 天梯 新增`user_rank`表记录用户的排名
-    await this.client.query('CREATE TABLE IF NOT EXISTS user_rank (id varchar not null, rank_value integer default 0, mu float4, sigma float4,trueskill float4 PRIMARY KEY (id))');
+    await this.client.query('CREATE TABLE IF NOT EXISTS user_rank (id varchar not null, rank_value integer default 0, mu float4, sigma float4,trueskill float4, PRIMARY KEY (id))');
     // 天梯 玩家数据表，用于保存段位的历史记录，和未来的数据分析 TODO: 未来如果做分析的话加上index
     await this.client.query('CREATE TABLE IF NOT EXISTS user_game_results (user_id varchar not null, game_id varchar not null, players integer, generations integer, createtime timestamp(0) default now(), corporation text, position integer, player_score integer, rank_value integer, mu float4, sigma float4,trueskill float4, is_rank integer, phase text, PRIMARY KEY (user_id, game_id))');
   }
@@ -77,11 +87,7 @@ export class PostgreSQL implements IDatabase {
   public async getGames(): Promise<Array<IGameShortData>> {
     const sql: string = 'SELECT games.game_id,games.prop FROM games, (SELECT max(save_id) save_id, game_id FROM games  GROUP BY game_id) a WHERE games.game_id = a.game_id AND games.save_id = a.save_id ORDER BY createtime DESC';
     const res = await this.client.query(sql);
-    return res.rows.map((row) => ({gameId: row.game_id, shortData: row.prop !== undefined && row.prop !=='' ? JSON.parse(row.prop) : undefined}));
-  }
-
-  public loadCloneableGame(gameId: GameId): Promise<SerializedGame> {
-    return this.getGameVersion(gameId, 0);
+    return res.rows.map((row) => ({gameId: row.game_id, shortData: row.prop !== undefined && row.prop !=='' ? row.prop : undefined}));
   }
 
   public async getGame(gameId: GameId): Promise<SerializedGame> {
@@ -147,7 +153,7 @@ export class PostgreSQL implements IDatabase {
 
   async cleanGame(gameId: GameId): Promise<void> {
     const maxSaveId = await this.getMaxSaveId(gameId);
-    console.log(`maxSaveId：${maxSaveId}, game_id:${gameId}  `);
+    console.log(`maxSaveId: ${maxSaveId}, game_id:${gameId}  `);
     // DELETE all saves except initial and last one
     await this.client.query('DELETE FROM games WHERE game_id = $1 AND save_id < $2 AND save_id > 0', [gameId, maxSaveId]);
     // Flag game as finished
@@ -156,20 +162,30 @@ export class PostgreSQL implements IDatabase {
     // const delete3 = this.purgeUnfinishedGames();
     // await Promise.all([delete1, delete2]);
   }
+  async markFinished(gameId: GameId): Promise<void> {
+    const promise1 = this.client.query('UPDATE games SET status = \'finished\' WHERE game_id = $1', [gameId]);
+    const promise2 = this.client.query('INSERT INTO completed_game(game_id) VALUES ($1)', [gameId]);
+    await Promise.all([promise1, promise2]);
+  }
 
   // Purge unfinished games older than MAX_GAME_DAYS days. If this environment variable is absent, it uses the default of 10 days.
-  async purgeUnfinishedGames(maxGameDays: string | undefined = process.env.MAX_GAME_DAYS): Promise<void> {
+  async purgeUnfinishedGames(maxGameDays: string | undefined = process.env.MAX_GAME_DAYS): Promise<Array<GameId>> {
     const dateToSeconds = daysAgoToSeconds(maxGameDays, 10);
     const selectResult = await this.client.query('SELECT DISTINCT game_id FROM games WHERE created_time < to_timestamp($1)', [dateToSeconds]);
-    const gameIds = selectResult.rows.slice(0, 1000).map((row) => row.game_id);
-    console.log(`${gameIds.length} games to be purged.`);
+    let gameIds = selectResult.rows.map((row) => row.game_id);
     if (gameIds.length > 1000) {
-      gameIds.length = 1000;
       console.log('Truncated purge to 1000 games.');
+      gameIds = gameIds.slice(0, 1000);
+    } else {
+      console.log(`${gameIds.length} games to be purged.`);
     }
-    // https://github.com/brianc/node-postgres/wiki/FAQ#11-how-do-i-build-a-where-foo-in--query-to-find-rows-matching-an-array-of-values
-    const deleteGamesResult = await this.client.query('DELETE FROM games WHERE game_id = ANY($1)', [gameIds]);
-    console.log(`Purged ${deleteGamesResult.rowCount} rows from games`);
+
+    if (gameIds.length > 0) {
+      // https://github.com/brianc/node-postgres/wiki/FAQ#11-how-do-i-build-a-where-foo-in--query-to-find-rows-matching-an-array-of-values
+      const deleteGamesResult = await this.client.query('DELETE FROM games WHERE game_id = ANY($1)', [gameIds]);
+      console.log(`Purged ${deleteGamesResult.rowCount} rows from games`);
+    }
+    return gameIds;
   }
 
   cleanGameAllSaves(game_id: string): void {
@@ -189,7 +205,7 @@ export class PostgreSQL implements IDatabase {
       }
     });
   }
-  async restoreGame(game_id: string, save_id: number, game: Game, playId: string): Promise<void> {
+  async restoreGame(game_id: string, save_id: number, game: IGame, playId: string): Promise<void> {
     // Retrieve last save from database
     logForUndo(game_id, 'restore to', save_id);
     const res = await this.client.query('SELECT game game ,createtime createtime  FROM games WHERE game_id = $1 AND save_id = $2 ORDER BY save_id DESC LIMIT 1', [game_id, save_id]);
@@ -207,21 +223,50 @@ export class PostgreSQL implements IDatabase {
     // Rebuild each objects
     const gamelog = game.gameLog;
     logForUndo(game.id, 'restored to', game.lastSaveId, 'from', save_id);
-    game.loadFromJSON(gameToRestore);
+    game.loadFromJSON(gameToRestore, true);
     game.gameLog = gamelog;
-    game.log('${0} undo turn', (b) => b.playerId(playId));
     // 这里undoCount取得是数据库中的值+1，对于连续的 一动-撤回-一动-撤回， 只会计算一次，但是没啥影响
     game.undoCount ++;
     // 会员回退时 以当前时间开始计时， 避免计时算到上一个人头上
     if (playId === 'manager') {
+      game.log('${0} undo turn', (b) => b.playerColor(playId as Color));
       Timer.newInstance().stop();
       game.activePlayer.timer.start();
+    } else {
+      game.log('${0} undo turn', (b) => b.player(game.getPlayerById(playId as PlayerId)));
     }
     console.log(`${playId} undo turn ${game_id}  ${save_id}`);
     return Promise.resolve();
   }
 
-  async saveGame(game: Game): Promise<void> {
+
+  async compressCompletedGames(compressCompletedGamesDays: string | undefined = process.env.COMPRESS_COMPLETED_GAMES_DAYS): Promise<void> {
+    if (compressCompletedGamesDays === undefined) {
+      return;
+    }
+    const dateToSeconds = daysAgoToSeconds(compressCompletedGamesDays, 0);
+    const selectResult = await this.client.query('SELECT DISTINCT game_id FROM completed_game WHERE completed_time < to_timestamp($1)', [dateToSeconds]);
+    const gameIds = selectResult.rows.slice(0, 1000).map((row) => row.game_id);
+    console.log(`${gameIds.length} completed games to be compressed.`);
+    if (gameIds.length > 1000) {
+      gameIds.length = 1000;
+      console.log('Compressing 1000 games.');
+    }
+    for (const gameId of gameIds) {
+      // This isn't using await because nothing really depends on it.
+      this.compressCompletedGame(gameId);
+    }
+  }
+
+  async compressCompletedGame(gameId: GameId): Promise<pg.QueryResult<any>> {
+    const maxSaveId = await this.getMaxSaveId(gameId);
+    return this.client.query('DELETE FROM games WHERE game_id = $1 AND save_id < $2 AND save_id > 0', [gameId, maxSaveId])
+      .then(() => {
+        return this.client.query('DELETE FROM completed_game where game_id = $1', [gameId]);
+      });
+  }
+
+  async saveGame(game: IGame): Promise<void> {
     try {
       const gameJSON = game.toJSON();
       const prop = game.toShortJSON();
@@ -278,9 +323,8 @@ export class PostgreSQL implements IDatabase {
     const res = await this.client.query('DELETE FROM games WHERE ctid IN (SELECT ctid FROM games WHERE game_id = $1 ORDER BY save_id DESC LIMIT $2)', [gameId, rollbackCount]);
     logForUndo(gameId, 'deleted', res?.rowCount, 'rows');
     const second = await this.getSaveIds(gameId);
-    const difference = first.filter((x) => !second.includes(x));
     logForUndo(gameId, 'second', second);
-    logForUndo(gameId, 'Rollback difference', difference);
+    logForUndo(gameId, 'Rollback difference', oneWayDifference(first, second));
   }
 
   public async storeParticipants(entry: GameIdLedger): Promise<void> {
@@ -290,7 +334,7 @@ export class PostgreSQL implements IDatabase {
   public async getParticipants(): Promise<Array<{gameId: GameId, participantIds: Array<ParticipantId>}>> {
     const res = await this.client.query('select game_id, participants from participants');
     return res.rows.map((row) => {
-      return {gameId: row.game_id as GameId, participantIds: row.participants as Array<ParticipantId>};
+      return {gameId: safeCast(row.game_id, isGameId), participantIds: row.participants as Array<ParticipantId>};
     });
   }
   saveUser(id: string, name: string, password: string, prop: string): void {
@@ -310,7 +354,7 @@ export class PostgreSQL implements IDatabase {
       }
       if (res && res.rows.length > 0) {
         res.rows.forEach((row) => {
-          const user = Object.assign(new User('', '', ''), {id: row.id, name: row.name, password: row.password, createtime: row.createtime}, JSON.parse(row.prop) );
+          const user = Object.assign(new User('', '', ''), {id: row.id, name: row.name, password: row.password, createtime: row.createtime}, row.prop );
           if (user.donateNum === 0 && user.isvip() > 0) {
             user.donateNum = 1;
           }
@@ -354,7 +398,7 @@ export class PostgreSQL implements IDatabase {
     // shows some bloat
     // and the postgres command
     // VACUUM (VERBOSE) shows a fairly reasonable vacumm (no rows locked, for instance),
-    // so it's not clear why those wrong. But these select count(*0) commands seem pretty quick
+    // so it's not clear why those wrong. But these select count(*) commands seem pretty quick
     // in testing. :fingers-crossed:
     for (const table of ['games', 'game_results', 'participants']) {
       const result = await this.client.query('select count(*) as rowcount from ' + table);
